+++ AzureServiceBusMessageTool/Main/Validations/IReceiverSettingsValidator.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Threading;
+using System.Threading.Tasks;
+using Core.Maybe;
+using Main.ViewModels.Configs.Receivers;
+
+namespace Main.Validations;
+
+public interface IReceiverSettingsValidator
+{
+    Task<Maybe<ValidationErrorResult>> Validate(ServiceBusReceiverSettings settings, CancellationToken token);
+}