+++ AzureServiceBusMessageTool/Main/ViewModels/Configs/Receivers/ReceivedMessage.cs
@@ -1,0 +1,6 @@
+﻿namespace Main.ViewModels.Configs.Receivers;
+
+public class ReceivedMessage
+{
+    public string Body { get; init; }
+}